+++ src/main/kotlin/cn/sicnu/postgraduate/core/service/PlanServiceImpl.kt
@@ -1,7 +1,6 @@
 package cn.sicnu.postgraduate.core.service
 
 import cn.hutool.core.convert.Convert
-import cn.hutool.core.date.LocalDateTimeUtil
 import cn.hutool.core.lang.Snowflake
 import cn.hutool.core.util.IdUtil
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper
@@ -16,11 +15,14 @@ import java.time.temporal.TemporalAdjusters
 import cn.sicnu.postgraduate.core.mapper.PlanMapper
 import cn.sicnu.postgraduate.core.entity.Plan
 import cn.sicnu.postgraduate.core.exception.CustomException
+import cn.sicnu.postgraduate.core.entity.User
+import cn.sicnu.postgraduate.springsecurity.entity.LoginUser
 import org.springframework.cache.CacheManager
 import org.springframework.cache.annotation.CacheEvict
 import org.springframework.cache.annotation.CachePut
 import org.springframework.cache.annotation.Cacheable
-import org.yaml.snakeyaml.events.Event.ID
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
+import org.springframework.security.core.context.SecurityContextHolder
 
 /*
     计划服务
@@ -43,13 +45,11 @@ class PlanServiceImpl(
     private val planMapper: PlanMapper,
     private val cacheManager: CacheManager
 ): PlanService {
-
     companion object {
         // 日志模块
         private val logger: Logger = LoggerFactory.getLogger(PlanService::class.java)
 
         // 常量声明
-        private const val CODE_SUCCESS: Int = 0
         private const val CODE_MISSING: Int = -1
         private const val MESSAGE_MISSING: String = "计划不存在"
 
@@ -59,10 +59,18 @@ class PlanServiceImpl(
         private const val MESSAGE_DATABASE_ERROR: String = "数据库错误"
         private const val CODE_WRONG_TIME: Int = -5
         private const val MESSAGE_WRONG_TIME: String = "时间参数错误"
+
+        private const val CODE_PERMISSION_DENIED: Int = -2
+        private const val MESSAGE_PERMISSION_DENIED: String = "无权限"
     }
 
+
+    /*  查询计划
+        pidStr      计划id
+     */
     @Cacheable(value = ["plan"], key = "#pidStr")
     override fun getPlan(pidStr: String): Plan {
+        //转换参数
         val pid: Long = Convert.toLong(pidStr)
 
         val plan: Plan? = planMapper.selectById(pid)
@@ -75,8 +83,14 @@ class PlanServiceImpl(
         }
     }
 
+    /*  批量查询计划
+        uidStr      用户id
+        lBeginDateStr       起始时间
+        lEndDateStr     结束时间
+     */
     @Cacheable(value = ["plan"], key = "#pidStr")
     override fun getPlanBy(uidStr: String, lBeginDateStr: String?, lEndDateStr: String?): List<Plan> {
+        //转换参数
         val uid: Long = Convert.toLong(uidStr)
         val lBeginDate: Long = Convert.toLong(lBeginDateStr)
         val lEndDate: Long = Convert.toLong(lEndDateStr)
@@ -125,22 +139,39 @@ class PlanServiceImpl(
         return plans
     }
 
+    /*  新建计划
+        uidStr      用户id
+        lDateStr        创建时间（timestamp milliseconds）
+        content     内容
+     */
     @CachePut(value = ["plan"], key = "T(java.lang.String).valueOf(#result.pid)")
     override fun newPlan(uidStr: String, lDateStr: String, content: String): Plan {
+        //转换参数
         val uid: Long = Convert.toLong(uidStr)
         val lDate: Long = Convert.toLong(lDateStr)
+
+        //鉴权
+        val user: User = getUser()
+        if(user.getUid() != uid) {
+            //鉴权失败，试图冒充他人
+            logger.error("newPlan: forwardUid: {},realUid: {}, {}", uid, user.getUid(), MESSAGE_PERMISSION_DENIED)
+            throw CustomException(CODE_PERMISSION_DENIED, MESSAGE_PERMISSION_DENIED)
+        }
 
         //转换时间
         var date: LocalDateTime = timestampConvert(lDate)
 
+        //生成新计划
         val snowFlake: Snowflake = IdUtil.getSnowflake()
         val pid: Long = snowFlake.nextId()
         val newPlan = Plan().apply {
+            setPid(pid)
             setUid(uid)
             setDate(date)
             setContent(content)
         }
 
+        //新建计划
         val result: Int = planMapper.insert(newPlan)
         if (result == 1) {
             return newPlan
@@ -155,8 +186,14 @@ class PlanServiceImpl(
         }
     }
 
+    /*  修改计划
+        pidStr      计划id
+        lDateStr        创建时间（timestamp milliseconds）
+        content     内容
+     */
     @CachePut(value = ["plan"], key = "#pidStr")
     override fun alterPlan(pidStr: String, lDateStr: String?, content: String?): Plan {
+        //参数转换
         val pid: Long = Convert.toLong(pidStr)
         val lDate: Long = Convert.toLong(lDateStr)
 
@@ -164,13 +201,24 @@ class PlanServiceImpl(
         var date: LocalDateTime? = null
         if(lDate != null){ date = timestampConvert(lDate) }
 
+        //未修改
         if (date == null && content == null) {
             logger.info("alterPlan: pid {}, {}", pid, MESSAGE_UNCHANGED)
             throw CustomException(CODE_UNCHANGED, MESSAGE_UNCHANGED)
         }
 
+        //修改计划
         val updatePlan: Plan? = planMapper.selectById(pid)
+
         if (updatePlan != null) {
+            //鉴权
+            val user: User = getUser()
+            if(user.getUid() != updatePlan.getUid()) {
+                //鉴权失败，试图修改他人计划
+                logger.error("alterPlan: pid {},uid: {}, {}", pid, user.getUid(), MESSAGE_PERMISSION_DENIED)
+                throw CustomException(CODE_PERMISSION_DENIED, MESSAGE_PERMISSION_DENIED)
+            }
+
             // 比对修改内容
             date?.let { updatePlan.setDate(it) }
             content?.let { updatePlan.setContent(it) }
@@ -195,12 +243,26 @@ class PlanServiceImpl(
         }
     }
 
+    /*  删除计划
+        pidStr      计划id
+     */
     @CacheEvict(value = ["plan"], key = "#pidStr")
     override fun deletePlan(pidStr: String): Plan {
+        //参数转换
         val pid: Long = Convert.toLong(pidStr)
 
+        //删除计划
         val dPlan: Plan? = planMapper.selectById(pid)
         if (dPlan != null) {
+            //鉴权
+            val user: User = getUser()
+            if(user.getUid() != dPlan.getUid()) {
+                //鉴权失败，试图删除他人计划
+                logger.error("deletePlan: pid {},uid: {}, {}", pid, user.getUid(), MESSAGE_PERMISSION_DENIED)
+                throw CustomException(CODE_PERMISSION_DENIED, MESSAGE_PERMISSION_DENIED)
+            }
+
+            //删除
             val result: Int = planMapper.deleteById(pid)
             if (result == 1) {
                 return dPlan
@@ -221,5 +283,12 @@ class PlanServiceImpl(
 
     private fun timestampConvert(lDate: Long): LocalDateTime {
         return LocalDateTime.ofInstant(Instant.ofEpochMilli(lDate), ZoneOffset.UTC)
+    }
+
+    private fun getUser(): User{
+        val auth: UsernamePasswordAuthenticationToken = SecurityContextHolder.getContext().getAuthentication() as UsernamePasswordAuthenticationToken
+        val loginUser: LoginUser = auth.getPrincipal() as LoginUser
+        val user: User = loginUser.getUser() ?: User()
+        return user
     }
 }
